+++ youtube_client_app/src/app/components/core/header/header.component.ts
