+++ youtube_client_app/src/app/components/core/header/profile/profile.component.ts
