+++ youtube_client_app/src/app/components/core/header/search-panel/search-panel.component.ts
